Adds delegate to determine if form data is logged
This adds a delegate that defaults to current behavior for determining if
a request's form data is logged (i.e. StatusCode == 500). This allows for
scenarios for someone to choose to log all form data for anything >= 400
except 404 errors for example.

It marks the LogPostedFormData as obsolete which might be heavy handed for
this release. It might make sense to just leave the remark pointing the
consumer towards the delegate now. Move it to obsolete for the next major
release and finally killing it off in the subsequent major
release...assuming of course there will be some major releases for this
package moving forward.