Reworks form logging to use OnMatch enum
Reworks the logging framework to use a fourth enumeration value that
indicates whether or not the custom predicate should be used. This allows
existing code to continue to function as is, but allows for a new option
for those that want to get a little fancier.

Also tweaked the test app to add a small test to verify that the should
log predicate functionality is working properly.